+++ lib/src/boxes/hivez_isolated_lazy.dart
@@ -5,6 +5,9 @@ import '../core/core.dart';
 
 class HivezIsolatedLazyBox<K, T>
     extends AbstractHivezIsolatedBox<K, T, IsolatedLazyBox<T>> {
+  @override
+  bool get isLazy => true;
+
   HivezIsolatedLazyBox(
     super.boxName, {
     super.encryptionCipher,
@@ -41,7 +44,7 @@ class HivezIsolatedLazyBox<K, T>
       final keys = (await box.keys).cast<K>();
       final List<T> values = [];
       for (final key in keys) {
-        final value = await box.get(key as T); // cast because keys are dynamic
+        final value = await box.get(key);
         if (value != null) values.add(value);
       }
       return values;